+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 ## 0.0.2 - unreleased
+### Breaking Changes
+- Change config load path to `server/data/config.toml` (previous: `server/config.toml`)
+
+
+### Other Changes
+**Technical**
+- deps(all): replace ts-result with better-result
 
 
 ## 0.0.1 - 2026-05-22